Move ActionCard into abstraction/action. Rename attack effect properties. Change select character signature. Bug fix in Attack class. Add getPlayer method for Effect interface. Add explanation in EffectFunctions. Remove Target since it's not an effect. Extend Character.attack method
