+++ src/CommunicationSpec.md
@@ -17,7 +17,7 @@ The available message_id are shown below:
 - [enter_room](#enter_room)
 - [enter_random](#enter_random)
 - [ready_start](#ready_start)
-- play_card
+- [play_card](#play_card)
 - move
 - end_round
 - surrender
@@ -70,3 +70,29 @@ Send when client is ready for a game
     "msg": "ready_start",
     "data": {}
 }
+```
+
+### play_card
+Rrerequest: game has started
+
+Send when client want to play a card
+
+```json
+{
+    "msg": "play_card",
+    "data": {
+        "card_id": integer
+    }
+}
+```
+
+<table>
+<tr>
+<th colspan="2">Fields</th>
+</tr>
+
+<tr>
+<td>card_id</td>
+<td>The id of the card the client want to play. For more information about card id, see <a href="#Card_id">Card_id</a> section.</td>
+</tr>
+</table>
