+++ src/main/java/eod/Player.java
@@ -5,14 +5,11 @@ import eod.IO.Output;
 import eod.card.abstraction.Card;
 import eod.card.abstraction.ICard;
 import eod.card.abstraction.action.ConditionalCard;
-import eod.effect.Target;
 import eod.listener.AttackListener;
 import eod.snapshots.Snapshotted;
 
 import java.awt.*;
 import java.util.*;
-
-import static eod.effect.EffectFunctions.Target;
 
 public class Player implements Snapshotted, GameObject {
 
@@ -155,7 +152,10 @@ public class Player implements Snapshotted, GameObject {
     }
 
     public void attack(Character from, Character[] to, int hp, boolean allowCondition, boolean willSuccess) {
-        Target targets = Target().on(to);
+        for(Character target:to) {
+            target.isTargeted = true;
+        }
+
         if(allowCondition) {
             game.triggerTargetedListener(this, willSuccess);
             to = Arrays.stream(to)
@@ -165,7 +165,9 @@ public class Player implements Snapshotted, GameObject {
         for(Character target:to) {
             target.damage(hp);
         }
-        targets.unTarget();
+        for(Character target:to) {
+            target.isTargeted = false;
+        }
     }
 
     public void targetedTrigger() {
