+++ src/main/java/eod/card/abstraction/action/ActionCard.java
@@ -1,6 +1,7 @@
-package eod.card.abstraction;
+package eod.card.abstraction.action;
 
 import eod.Player;
+import eod.card.abstraction.Card;
 
 public abstract class ActionCard extends Card {
     public final int level;
