+++ src/main/java/eod/card/abstraction/action/AttackCard.java
@@ -1,7 +1,6 @@
 package eod.card.abstraction.action;
 
 import eod.Player;
-import eod.card.abstraction.ActionCard;
 
 public abstract class AttackCard extends ActionCard {
     public Player rival;
