+++ src/main/java/eod/card/abstraction/action/ConditionalCard.java
@@ -1,7 +1,6 @@
 package eod.card.abstraction.action;
 
 import eod.Player;
-import eod.card.abstraction.ActionCard;
 
 public abstract class ConditionalCard extends ActionCard {
     public ConditionalCard(Player p, int lv) {
