+++ src/main/java/eod/card/abstraction/action/NormalCard.java
@@ -1,7 +1,6 @@
 package eod.card.abstraction.action;
 
 import eod.Player;
-import eod.card.abstraction.ActionCard;
 
 public abstract class NormalCard extends ActionCard {
     public NormalCard(Player p, int lv) {
